ml-service/requirements.txt
numpy
opencv-python==4.5.5.64
opencv-contrib-python==4.5.5.64
pytesseract
fastapi
uvicorn
scikit-learn==0.24.2
python-multipart
pillow
